refactor: Split admin.go, typed constraint errors, batched cleanup, safe defaults
Split admin.go into per-group files:
- admin.go (dispatch + shared helpers)
- admin_user.go, admin_session.go, admin_worker.go
- admin_oauth.go, admin_encryption.go, admin_db.go

friendlyConstraintError now uses SQLite error codes
(errors.As + sqlite.Error.Code() == 2067) instead of brittle
substring matching, falling back to message inspection only for
UNIQUE constraint violations.

Cleanup loop batching:
- All hard-delete queries use LIMIT 1000 via rowid subquery
- step/cleanupStep helpers loop until 0 rows affected
- Prevents long write-lock holds on bulk deletions

Safe defaults for GetUserByID / GetOrgByID:
- GetUserByID and GetOrgByID now filter deleted_at IS NULL
- New GetUserByIDIncludeDeleted / GetOrgByIDIncludeDeleted for
  admin and test code that needs to see soft-deleted records
- Removed redundant manual DeletedAt.Valid check in GetOrg RPC

Other:
- Renamed ForceDeleteOrg to SoftDeleteOrg for clarity
- Refactored createTestSession/createTestWorker to accept
  *gendb.Queries instead of reopening the DB
- Added hub cleanup integration tests (TestRun_CleansUpOldRecords,
  TestRun_RetainsRecentlyDeleted)
