+++ backend/internal/hub/db/queries/orgs.sql
@@ -2,6 +2,9 @@
 INSERT INTO orgs (id, name, is_personal) VALUES (?, ?, ?);
 
 -- name: GetOrgByID :one
+SELECT * FROM orgs WHERE id = ? AND deleted_at IS NULL;
+
+-- name: GetOrgByIDIncludeDeleted :one
 SELECT * FROM orgs WHERE id = ?;
 
 -- name: GetOrgByName :one
@@ -16,8 +19,8 @@ UPDATE orgs SET name = ? WHERE id = ? AND is_personal = 0;
 -- name: DeleteOrg :exec
 UPDATE orgs SET deleted_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now') WHERE id = ? AND is_personal = 0;
 
--- name: ForceDeleteOrg :exec
+-- name: SoftDeleteOrg :exec
 UPDATE orgs SET deleted_at = strftime('%Y-%m-%dT%H:%M:%fZ', 'now') WHERE id = ?;
 
 -- name: HardDeleteOrgsBefore :execresult
-DELETE FROM orgs WHERE deleted_at IS NOT NULL AND deleted_at < ?;
+DELETE FROM orgs WHERE rowid IN (SELECT o.rowid FROM orgs o WHERE o.deleted_at IS NOT NULL AND o.deleted_at < ? LIMIT 1000);
